fix: bad type assignation

+++ main.py
@@ -10,7 +10,7 @@ class ArgsSchema:
     action: str
     description: str | None = field(default=None)
     id: int | None = field(default=None)
-    amount: int | None = field(default=None)
+    amount: float | None = field(default=None)
     month: int | None = field(default=None)
